GCC and clang working?

+++ libraries/imgui/src/imgui_layer.cpp
@@ -20,8 +20,8 @@ void ImGuiLayer::on_attach()
     IMGUI_CHECKVERSION();
     ImGui::CreateContext();
     ImGuiIO& imgui_io = ImGui::GetIO();
-    imgui_io.ConfigFlags |= static_cast<uint32_t>(ImGuiConfigFlags_NavEnableKeyboard); // Enable Keyboard Controls
-    imgui_io.ConfigFlags |= static_cast<uint32_t>(ImGuiConfigFlags_NavEnableGamepad);  // Enable Gamepad Controls
+    imgui_io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard; // Enable Keyboard Controls
+    imgui_io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;  // Enable Gamepad Controls
 
     ImGui::StyleColorsDark();
